Remove unused functions for Line
