+++ crates/geometry/src/line.rs
@@ -6,7 +6,7 @@ use serde::{Deserialize, Serialize};
 use linear_algebra::{distance_squared, point, vector, Point, Point2, Transform};
 use path_serde::{PathDeserialize, PathIntrospect, PathSerialize};
 
-use crate::{line_segment::LineSegment, signed_acute_angle, Distance};
+use crate::{line_segment::LineSegment, Distance};
 
 #[derive(
     Copy,
@@ -28,27 +28,6 @@ pub type Line2<Frame> = Line<Frame, 2>;
 pub type Line3<Frame> = Line<Frame, 3>;
 
 impl<Frame> Line2<Frame> {
-    pub fn signed_acute_angle(&self, other: Self) -> f32 {
-        let self_direction = self.1 - self.0;
-        let other_direction = other.1 - other.0;
-        signed_acute_angle(self_direction, other_direction)
-    }
-
-    pub fn angle(&self, other: Self) -> f32 {
-        (self.1 - self.0).angle(other.1 - other.0)
-    }
-
-    pub fn signed_acute_angle_to_orthogonal(&self, other: Self) -> f32 {
-        let self_direction = self.1 - self.0;
-        let other_direction = other.1 - other.0;
-        let orthogonal_other_direction = vector![other_direction.y(), -other_direction.x()];
-        signed_acute_angle(self_direction, orthogonal_other_direction)
-    }
-
-    pub fn is_orthogonal(&self, other: Self, epsilon: f32) -> bool {
-        self.signed_acute_angle_to_orthogonal(other).abs() < epsilon
-    }
-
     pub fn slope(&self) -> f32 {
         let difference = self.0 - self.1;
         difference.y() / difference.x()
@@ -67,19 +46,6 @@ impl<Frame> Line2<Frame> {
         let line_vector = self.1 - self.0;
         let normal_vector = vector![-line_vector.y(), line_vector.x()].normalize();
         normal_vector.dot(point.coords()) - normal_vector.dot(self.0.coords())
-    }
-
-    pub fn project_onto_segment(&self, point: Point2<Frame>) -> Point2<Frame> {
-        let difference_on_line = self.1 - self.0;
-        let difference_to_point = point - self.0;
-        let t = difference_to_point.dot(difference_on_line) / difference_on_line.norm_squared();
-        if t <= 0.0 {
-            self.0
-        } else if t >= 1.0 {
-            self.1
-        } else {
-            self.0 + difference_on_line * t
-        }
     }
 
     pub fn intersection(&self, other: &Line2<Frame>) -> Point2<Frame> {
@@ -167,135 +133,5 @@ impl<Frame, const DIMENSION: usize> RelativeEq for Line<Frame, DIMENSION> {
 impl<Frame> From<LineSegment<Frame>> for Line2<Frame> {
     fn from(line_segment: LineSegment<Frame>) -> Self {
         Self(line_segment.0, line_segment.1)
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::f32::consts::FRAC_PI_4;
-
-    use approx::assert_relative_eq;
-
-    use super::*;
-
-    #[derive(Clone, Copy, Debug)]
-    struct SomeFrame;
-
-    #[test]
-    fn correct_acute_signed_angle() {
-        #[derive(Debug)]
-        struct Case {
-            self_line: Line2<SomeFrame>,
-            other_line: Line2<SomeFrame>,
-            expected_angle: f32,
-        }
-
-        let thirty_degree = 30.0_f32.to_radians();
-        let sixty_degree = 60.0_f32.to_radians();
-        let cases = [
-            Case {
-                self_line: Line(point![0.0, 0.0], point![42.0, 0.0]),
-                other_line: Line(point![0.0, 0.0], point![42.0, 0.0]),
-                expected_angle: 0.0,
-            },
-            Case {
-                self_line: Line(point![0.0, 0.0], point![42.0, 0.0]),
-                other_line: Line(point![0.0, 0.0], point![42.0, 42.0]),
-                expected_angle: FRAC_PI_4,
-            },
-            Case {
-                self_line: Line(point![0.0, 0.0], point![42.0, 42.0]),
-                other_line: Line(point![0.0, 0.0], point![42.0, 0.0]),
-                expected_angle: -FRAC_PI_4,
-            },
-            Case {
-                self_line: Line(point![0.0, 0.0], point![42.0, 0.0]),
-                other_line: Line(point![0.0, 0.0], point![42.0, -42.0]),
-                expected_angle: -FRAC_PI_4,
-            },
-            Case {
-                self_line: Line(point![0.0, 0.0], point![42.0, -42.0]),
-                other_line: Line(point![0.0, 0.0], point![42.0, 0.0]),
-                expected_angle: FRAC_PI_4,
-            },
-            Case {
-                self_line: Line(
-                    point![0.0, 0.0],
-                    point![(-thirty_degree).cos(), (-thirty_degree).sin()],
-                ),
-                other_line: Line(
-                    point![0.0, 0.0],
-                    point![thirty_degree.cos(), thirty_degree.sin()],
-                ),
-                expected_angle: sixty_degree,
-            },
-            Case {
-                self_line: Line(
-                    point![0.0, 0.0],
-                    point![thirty_degree.cos(), thirty_degree.sin()],
-                ),
-                other_line: Line(
-                    point![0.0, 0.0],
-                    point![(-thirty_degree).cos(), (-thirty_degree).sin()],
-                ),
-                expected_angle: -sixty_degree,
-            },
-            Case {
-                self_line: Line(
-                    point![0.0, 0.0],
-                    point![(-sixty_degree).cos(), (-sixty_degree).sin()],
-                ),
-                other_line: Line(
-                    point![0.0, 0.0],
-                    point![sixty_degree.cos(), sixty_degree.sin()],
-                ),
-                expected_angle: -sixty_degree,
-            },
-            Case {
-                self_line: Line(
-                    point![0.0, 0.0],
-                    point![sixty_degree.cos(), sixty_degree.sin()],
-                ),
-                other_line: Line(
-                    point![0.0, 0.0],
-                    point![(-sixty_degree).cos(), (-sixty_degree).sin()],
-                ),
-                expected_angle: sixty_degree,
-            },
-        ]
-        .into_iter()
-        .flat_map(|case| {
-            [
-                Case {
-                    self_line: case.self_line,
-                    other_line: case.other_line,
-                    expected_angle: case.expected_angle,
-                },
-                Case {
-                    self_line: Line(case.self_line.1, case.self_line.0),
-                    other_line: case.other_line,
-                    expected_angle: case.expected_angle,
-                },
-                Case {
-                    self_line: case.self_line,
-                    other_line: Line(case.other_line.1, case.other_line.0),
-                    expected_angle: case.expected_angle,
-                },
-                Case {
-                    self_line: Line(case.self_line.1, case.self_line.0),
-                    other_line: Line(case.other_line.1, case.other_line.0),
-                    expected_angle: case.expected_angle,
-                },
-            ]
-        });
-
-        for case in cases {
-            dbg!(&case);
-            assert_relative_eq!(
-                case.self_line.signed_acute_angle(case.other_line),
-                case.expected_angle,
-                epsilon = 0.000001,
-            );
-        }
     }
 }
